% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_distribution.R
\name{check_fire_data}
\alias{check_fire_data}
\title{Check fire data for sufficiency}
\usage{
check_fire_data(fires_hist_size, sim_perimeters_size, n_years)
}
\arguments{
\item{fires_hist_size}{Numeric vector. A vector of historical fire sizes (e.g., areas in hectares).}

\item{sim_perimeters_size}{Numeric vector. A vector of simulated fire perimeter sizes (e.g., areas in hectares).}

\item{n_years}{Integer. The number of years represented by the historical fire data.}
}
\value{
If simulated data is sufficient, returns an integer representing a
  recommended surface threshold (typically 10% of the maximum possible threshold).
  Otherwise, returns `NULL` (invisibly, as it only prints messages) if
  simulated data is deemed insufficient.
}
\description{
This function assesses whether the available simulated fire perimeters are
sufficient in terms of maximum fire size and total burned area, relative
to historical fire data. It provides messages to the user if deficiencies
are found and, if data is sufficient, calculates and recommends a surface
threshold for selecting events.
}
\examples{
# Example 1: Simulated fires are too small and not enough area
check_fire_data(fires_hist_size = c(100, 500, 1000, 2000),
                sim_perimeters_size = c(10, 50, 80, 120),
                n_years = 1)

# Example 2: Simulated fires are sufficient
check_fire_data(fires_hist_size = c(100, 500, 1000, 2000),
                sim_perimeters_size = c(10, 50, 80, 120, 500, 1500, 2500, 5000),
                n_years = 1)

# Example 3: Insufficient area, but max size is okay
check_fire_data(fires_hist_size = c(100, 500, 1000, 2000),
                sim_perimeters_size = c(10, 50, 80, 120, 500, 1500),
                n_years = 10)
# Example 4: Correct data
check_fire_data(fires_hist_size = c(1, 5, 10, 600),
               sim_perimeters_size = c(10, 50, 80, 120, 500, 2500),
               n_years = 10)
}
\seealso{
\code{\link{get_select_params}}
}
