% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_select_events.R
\name{get_select_params}
\alias{get_select_params}
\title{Get Selection Parameters}
\usage{
get_select_params(fires_hist_size, sim_perimeters_size, n_years)
}
\arguments{
\item{fires_hist_size}{Numeric vector. A vector of historical fire sizes.}

\item{sim_perimeters_size}{Numeric vector. A vector of simulated fire perimeter sizes.}

\item{n_years}{Integer. The number of years covered by the historical fire data.}
}
\value{
An integer representing the ratio of total simulated area to the
  annual average historical area. This value can be interpreted as the
  maximum number of historical fire "equivalents" that can be formed from
  the simulated data.
}
\description{
Calculates a parameter for event selection based on the total area of
historical fires (normalized by number of years) and the total area of
simulated fire perimeters. This parameter can be used to determine a
maximum surface threshold for selecting a subset of simulated events.
}
\examples{
# Calculate selection parameter for a scenario where simulated data
# covers 10 times the annual historical area.
get_select_params(fires_hist_size = c(100, 500, 1000),
                  sim_perimeters_size = c(1000, 2000, 3000, 4000, 5000),
                  n_years = 1)

# If historical data represents 10 years
get_select_params(fires_hist_size = c(100, 500, 1000),
                  sim_perimeters_size = c(1000, 2000, 3000, 4000, 5000),
                  n_years = 10)
}
\seealso{
\code{\link{check_fire_data}}
}
