% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_distribution.R
\name{visualize_selected_dist}
\alias{visualize_selected_dist}
\title{Visualize Selected vs. Target Fire Size Distribution}
\usage{
visualize_selected_dist(
  result = result,
  logaritmic = TRUE,
  target_hist = target_hist,
  bins = bins
)
}
\arguments{
\item{result}{Lista devuelta por un modelo o proceso de selección
que debe contener al menos:
\describe{
  \item{selected_surfaces}{Vector numérico con los tamaños de superficie seleccionados.}
  \item{total_surface}{Valor numérico con la superficie total seleccionada.}
  \item{final_discrepancy}{Valor numérico con la discrepancia final respecto a la distribución objetivo.}
}}

\item{logaritmic}{Lógico. Si \code{TRUE}, los tamaños seleccionados se transforman logarítmicamente
antes de graficar. Por defecto \code{TRUE}.}

\item{target_hist}{Vector numérico con las densidades de la distribución objetivo,
correspondiente a los puntos medios de los intervalos definidos en \code{bins}.}

\item{bins}{Vector numérico con los límites de los intervalos (breaks) usados
para construir el histograma.}
}
\value{
Un objeto \code{ggplot} que representa la comparación entre la distribución
seleccionada y la distribución objetivo. Además, la función imprime en consola:
\itemize{
  \item Número de eventos seleccionados.
  \item Superficie total seleccionada.
  \item Discrepancia respecto a la distribución objetivo.
}
Si no se encuentra una solución, devuelve un mensaje y no genera gráfico.
}
\description{
Esta función genera un gráfico comparando la distribución de los
tamaños de incendio seleccionados con una distribución objetivo.
Permite aplicar transformación logarítmica a los tamaños seleccionados
y visualizar ambas distribuciones de manera conjunta.
}
\examples{
\dontrun{
# Use example for select_events
visualize_selected_dist(result, logaritmic = TRUE,
                        target_hist = target_hist, bins = bins)
}

}
